add integration utilities
